+++ changelog.md
@@ -7,9 +7,11 @@
 * Significantly refactor connection to lobby server (#620, #621)
 * Remove redundant APPDATA_DIR loading in util (#665, #666)
 * Fix theme breakage caused by a promoted widget (#669, #670)
+* Fix random/nomads faction shown in replay info (#676)
 
 Contributors:
   - Wesmania
+  - Grothe
 
 0.12.4
 =====